feat: 브랜드 목록 조회 /common 으로 변경
