+++ src/main/java/com/earth/ureverse/global/common/controller/CommonController.java
@@ -1,21 +1,30 @@
 package com.earth.ureverse.global.common.controller;
 
+import com.earth.ureverse.global.common.dto.response.BrandResponseDto;
 import com.earth.ureverse.global.common.response.CommonResponseEntity;
+import com.earth.ureverse.global.common.service.CommonService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.List;
+
 @Tag(name = "Common", description = "공통 응답 포맷의 예제를 제공합니다. 성공 및 에러 응답 구조를 확인할 수 있습니다.")
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/api/v1/common")
 public class CommonController {
+
+    private final CommonService commonService;
+
     //성공
     @Operation(summary = "성공 예시", description = "성공 응답 예제입니다.")
     @ApiResponse(responseCode = "200", description = "성공 응답",
@@ -35,5 +44,11 @@ public class CommonController {
         CommonResponseEntity<Object> body =
                 CommonResponseEntity.error(HttpStatus.INTERNAL_SERVER_ERROR, "An error occurred");
         return new ResponseEntity<>(body, HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+    @GetMapping("/allBrands")
+    public CommonResponseEntity<List<BrandResponseDto>> getAllBrand(){
+        List<BrandResponseDto> brandResponseDtos = commonService.getBrands();
+        return CommonResponseEntity.success(brandResponseDtos);
     }
 }
