+++ src/main/java/com/earth/ureverse/global/common/dto/response/BrandResponseDto.java
@@ -1,9 +1,8 @@
-package com.earth.ureverse.member.dto.response;
+package com.earth.ureverse.global.common.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
 @NoArgsConstructor
