+++ src/main/java/com/earth/ureverse/global/common/mapper/BrandMapper.java
@@ -1,12 +1,11 @@
-package com.earth.ureverse.member.mapper;
+package com.earth.ureverse.global.common.mapper;
 
-import com.earth.ureverse.member.dto.response.BrandResponseDto;
+import com.earth.ureverse.global.common.dto.response.BrandResponseDto;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
 
 @Mapper
 public interface BrandMapper {
-
     List<BrandResponseDto> selectAllBrands();
 }
