+++ src/main/java/com/earth/ureverse/global/common/service/CommonService.java
@@ -1,0 +1,10 @@
+package com.earth.ureverse.global.common.service;
+
+import com.earth.ureverse.global.common.dto.response.BrandResponseDto;
+
+import java.util.List;
+
+public interface CommonService {
+
+    List<BrandResponseDto> getBrands();
+}
