+++ src/main/java/com/earth/ureverse/member/service/MemberService.java
@@ -3,13 +3,10 @@ package com.earth.ureverse.member.service;
 import com.earth.ureverse.member.dto.request.ChangePasswordRequestDto;
 import com.earth.ureverse.member.dto.request.UpdateMemberRequestDto;
 import com.earth.ureverse.member.dto.request.WithdrawRequestDto;
-import com.earth.ureverse.member.dto.response.BrandResponseDto;
 import com.earth.ureverse.member.dto.response.MemberInfoResponseDto;
 import com.earth.ureverse.member.dto.response.PointHistoryListResponseDto;
 import com.earth.ureverse.member.dto.response.SalesHistoryListResponseDto;
 import jakarta.validation.Valid;
-
-import java.util.List;
 
 public interface MemberService {
 
@@ -24,7 +21,5 @@ public interface MemberService {
     SalesHistoryListResponseDto getSalesHistory(Long userId, int limit, String lastCreatedAt, Long lastProductId);
 
     MemberInfoResponseDto getMyInfo(Long userId);
-
-    List<BrandResponseDto> getBrands();
 
 }
